feat: I want the reset filter option on embedded dashboards
